Fix upgrade with aggregate functions for pg11
The upgrade of postgis_sfcgal extension on pg11 from 3.2.4 to 3.3.2
fails because the aggregate function is never created. It is unclear how
the generated SQL worked before but now it properly checks that the
function was introduced after the `to` version instead of the `from`
version. This allows the extension to be upgraded properly.